pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
authors = [{ name = "Mikhail Kokuev" }]
description = "Compares two configuration files and shows a difference."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.7",
]
